convert to float

+++ ionchannel_convbert_hyperparam.py
@@ -67,7 +67,7 @@ def create_datasets(
             label = self.labels[idx]
             # Convert string label to integer
             label_int = label_encoder.transform([label])[0]
-            label_int = torch.tensor(label_int).unsqueeze(-1)
+            label_int = torch.tensor(label_int, dtype=torch.float).unsqueeze(-1)
 
             return {
                 "embed": torch.tensor(embedding),
